Refactor storage to separate file, fixes
