+++ daemons/eth.py
@@ -1,13 +1,10 @@
 import asyncio
-import copy
 import functools
 import inspect
 import json
 import os
 import random
 import secrets
-import stat
-import threading
 import time
 import traceback
 from dataclasses import dataclass, field
@@ -20,6 +17,9 @@ from eth_account.messages import encode_defunct
 from eth_keys.datatypes import PrivateKey, PublicKey
 from hexbytes import HexBytes
 from mnemonic import Mnemonic
+from storage import JSONEncoder as StorageJSONEncoder
+from storage import Storage, StoredObject
+from storage import WalletDB as StorageWalletDB
 from utils import JsonResponse, get_exception_message, hide_logging_errors, rpc
 from web3 import Web3
 from web3.datastructures import AttributeDict
@@ -64,7 +64,23 @@ STR_TO_BOOL_MAPPING = {
 
 # storage
 
-WALLET_VERSION = 1
+
+class WalletDB(StorageWalletDB):
+    WALLET_VERSION = 1
+
+    def _convert_dict(self, path, key, v):
+        if key == "payment_requests":
+            v = {k: Invoice(**x) for k, x in v.items()}
+        return v
+
+
+def user_dir():
+    if os.name == "posix":
+        return os.path.join(os.environ["HOME"], ".bitcart-eth")
+    elif "APPDATA" in os.environ:
+        return os.path.join(os.environ["APPDATA"], "Bitcart-ETH")
+    elif "LOCALAPPDATA" in os.environ:
+        return os.path.join(os.environ["LOCALAPPDATA"], "Bitcart-ETH")
 
 
 def str_to_bool(s):
@@ -81,16 +97,12 @@ def is_address(address):
     return Web3.isAddress(address) or Web3.isChecksumAddress(address)
 
 
-class JSONEncoder(json.JSONEncoder):
+class JSONEncoder(StorageJSONEncoder):
     def default(self, obj):
         if isinstance(obj, AttributeDict):
             return {k: v for k, v in obj.items()}
         if isinstance(obj, HexBytes):
             return str(obj.hex())
-        if isinstance(obj, Decimal):
-            return str(obj)
-        if hasattr(obj, "to_json") and callable(obj.to_json):
-            return obj.to_json()
         return super().default(obj)
 
 
@@ -100,289 +112,6 @@ def to_dict(obj):
 
 def get_exception_traceback(exc):
     return "".join(traceback.format_exception(type(exc), exc, exc.__traceback__))
-
-
-def standardize_path(path):
-    return os.path.normcase(os.path.realpath(os.path.abspath(os.path.expanduser(path))))
-
-
-def user_dir():
-    if os.name == "posix":
-        return os.path.join(os.environ["HOME"], ".bitcart-eth")
-    elif "APPDATA" in os.environ:
-        return os.path.join(os.environ["APPDATA"], "Bitcart-ETH")
-    elif "LOCALAPPDATA" in os.environ:
-        return os.path.join(os.environ["LOCALAPPDATA"], "Bitcart-ETH")
-
-
-class WalletFileException(Exception):
-    pass
-
-
-class Storage:
-    def __init__(self, path):
-        self.path = standardize_path(path)
-        self._file_exists = bool(self.path and os.path.exists(self.path))
-        if self.file_exists():
-            with open(self.path, encoding="utf-8") as f:
-                self.raw = f.read()
-        else:
-            self.raw = ""
-
-    def read(self):
-        return self.raw
-
-    def write(self, data: str) -> None:
-        # write in temporary first to not corrupt the main file
-        s = data
-        temp_path = f"{self.path}.tmp.{os.getpid()}"
-        with open(temp_path, "w", encoding="utf-8") as f:
-            f.write(s)
-            f.flush()
-            os.fsync(f.fileno())
-        try:
-            mode = os.stat(self.path).st_mode
-        except FileNotFoundError:
-            mode = stat.S_IREAD | stat.S_IWRITE
-        if not self.file_exists():
-            assert not os.path.exists(self.path)
-        os.replace(temp_path, self.path)
-        os.chmod(self.path, mode)
-        self._file_exists = True
-
-    def file_exists(self) -> bool:
-        return self._file_exists
-
-
-def modifier(func):
-    def wrapper(self, *args, **kwargs):
-        with self.lock:
-            self._modified = True
-            return func(self, *args, **kwargs)
-
-    return wrapper
-
-
-def locked(func):
-    def wrapper(self, *args, **kwargs):
-        with self.lock:
-            return func(self, *args, **kwargs)
-
-    return wrapper
-
-
-class JsonDB:
-    def __init__(self, data):
-        self.lock = threading.RLock()
-        self.data = data
-        self._modified = False
-
-    def set_modified(self, b):
-        with self.lock:
-            self._modified = b
-
-    def modified(self):
-        return self._modified
-
-    @locked
-    def get(self, key, default=None):
-        v = self.data.get(key)
-        if v is None:
-            v = default
-        return v
-
-    @modifier
-    def put(self, key, value):
-        try:
-            json.dumps(key, cls=JSONEncoder)
-            json.dumps(value, cls=JSONEncoder)
-        except Exception:
-            return False
-        if value is not None:
-            if self.data.get(key) != value:
-                self.data[key] = copy.deepcopy(value)
-                return True
-        elif key in self.data:
-            self.data.pop(key)
-            return True
-        return False
-
-    @locked
-    def dump(self) -> str:
-        return json.dumps(self.data, cls=JSONEncoder)
-
-    def _should_convert_to_stored_dict(self, key) -> bool:
-        return True
-
-
-class StoredObject:
-
-    db = None
-
-    def __setattr__(self, key, value):
-        if self.db:
-            self.db.set_modified(True)
-        super().__setattr__(key, value)
-
-    def set_db(self, db):
-        self.db = db
-
-    def to_json(self):
-        d = dict(vars(self))
-        d.pop("db", None)
-        d = {k: v for k, v in d.items() if not k.startswith("_")}
-        return d
-
-
-_RaiseKeyError = object()
-
-
-class StoredDict(dict):
-    def __init__(self, data, db, path):
-        self.db = db
-        self.lock = self.db.lock if self.db else threading.RLock()
-        self.path = path
-        # recursively convert dicts to StoredDict
-        for k, v in list(data.items()):
-            self.__setitem__(k, v)
-
-    @locked
-    def __setitem__(self, key, v):
-        is_new = key not in self
-        # early return to prevent unnecessary disk writes
-        if not is_new and self[key] == v:
-            return
-        # recursively set db and path
-        if isinstance(v, StoredDict):
-            v.db = self.db
-            v.path = self.path + [key]
-            for k, vv in v.items():
-                v[k] = vv
-        # recursively convert dict to StoredDict.
-        # _convert_dict is called breadth-first
-        elif isinstance(v, dict):
-            if self.db:
-                v = self.db._convert_dict(self.path, key, v)
-            if not self.db or self.db._should_convert_to_stored_dict(key):
-                v = StoredDict(v, self.db, self.path + [key])
-        # set parent of StoredObject
-        if isinstance(v, StoredObject):
-            v.set_db(self.db)
-        # set item
-        super().__setitem__(key, v)
-        if self.db:
-            self.db.set_modified(True)
-
-    @locked
-    def __delitem__(self, key):
-        super().__delitem__(key)
-        if self.db:
-            self.db.set_modified(True)
-
-    @locked
-    def pop(self, key, v=_RaiseKeyError):
-        if v is _RaiseKeyError:
-            r = super().pop(key)
-        else:
-            r = super().pop(key, v)
-        if self.db:
-            self.db.set_modified(True)
-        return r
-
-    @locked
-    def clear(self):
-        super().clear()
-        if self.db:
-            self.db.set_modified(True)
-
-
-# Thanks to https://github.com/spesmilo/electrum/blob/master/electrum/wallet_db.py
-class WalletDB(JsonDB):
-    def __init__(self, raw):
-        super().__init__({})
-        self.upgraded = False
-        if raw:
-            self.load_data(raw)
-        else:
-            self.put("version", WALLET_VERSION)
-        self._after_upgrade_tasks()
-
-    def load_data(self, s):
-        try:
-            self.data = json.loads(s)
-        except Exception as e:
-            raise WalletFileException("Cannot read wallet file. (parsing failed)") from e
-        if not isinstance(self.data, dict):
-            raise WalletFileException("Malformed wallet file (not dict)")
-        if self.requires_upgrade():
-            self.upgrade()
-
-    def requires_upgrade(self):
-        return self.get_version() < WALLET_VERSION
-
-    def upgrade(self):
-        # future upgrade code here
-        self.put("version", WALLET_VERSION)
-        self._after_upgrade_tasks()
-
-    def _after_upgrade_tasks(self):
-        self.upgraded = True
-        self.data = StoredDict(self.data, self, [])
-
-    def _is_upgrade_method_needed(self, min_version, max_version):
-        assert min_version <= max_version
-        cur_version = self.get_version()
-        if cur_version > max_version:
-            return False
-        elif cur_version < min_version:
-            raise WalletFileException(
-                f"storage upgrade: unexpected version {cur_version} (should be {min_version}-{max_version})"
-            )
-        else:
-            return True
-
-    @locked
-    def get_version(self):
-        version = self.get("version")
-        if not version:
-            version = WALLET_VERSION
-        if version > WALLET_VERSION:
-            raise WalletFileException(
-                "This version of BitcartCC ETH daemon is too old to open this wallet.\n"
-                f"(highest supported storage version: {WALLET_VERSION}, version of this file: {version})"
-            )
-        return version
-
-    @locked
-    def get_dict(self, name) -> dict:
-        # Warning: interacts un-intuitively with 'put': certain parts
-        # of 'data' will have pointers saved as separate variables.
-        if name not in self.data:
-            self.data[name] = {}
-        return self.data[name]
-
-    def _convert_dict(self, path, key, v):
-        if key == "payment_requests":
-            v = {k: Invoice(**x) for k, x in v.items()}
-        return v
-
-    def _should_convert_to_stored_dict(self, key) -> bool:
-        if key == "keystore":
-            return False
-        return True
-
-    def write(self, storage):
-        with self.lock:
-            self._write(storage)
-
-    def _write(self, storage):
-        if not self.modified():
-            return
-        storage.write(self.dump())
-        self.set_modified(False)
-
-    def is_ready_to_be_used_by_wallet(self):
-        return not self.requires_upgrade() and self.upgraded
 
 
 @dataclass
@@ -548,10 +277,11 @@ class Wallet:
                 add_high *= 2
         return cur_amount
 
-    def add_payment_request(self, req):
+    def add_payment_request(self, req, save_db=True):
         self.receive_requests[req.id] = req
         self.used_amounts[str(req.amount)] = req.id
-        self.save_db()
+        if save_db:
+            self.save_db()
         return req
 
     async def get_request_url(self, req):
@@ -610,9 +340,10 @@ class Wallet:
         req.status = status
         for kwarg in kwargs:
             setattr(req, kwarg, kwargs[kwarg])
-        self.add_payment_request(req)
+        self.add_payment_request(req, save_db=False)
         if status != PR_UNPAID:
             self.used_amounts.pop(str(req.amount), None)
+        self.save_db()
         return req
 
     async def expired_task(self, req):
@@ -700,7 +431,7 @@ class ETHDaemon(BaseDaemon):
         wallet = self.addresses[to]
         tx_hash = str(tx["hash"].hex())
         await self.trigger_event({"event": "new_transaction", "tx": tx_hash}, wallet)
-        if amount in self.wallets[wallet].used_amounts:
+        if str(amount) in self.wallets[wallet].used_amounts:
             self.loop.create_task(self.process_payment(wallet, amount, tx_hash))
 
     async def process_pending(self):
@@ -729,10 +460,15 @@ class ETHDaemon(BaseDaemon):
                 self.wallets_updates[key].append(data)
 
     async def process_payment(self, wallet, amount, tx_hash):
-        req = self.wallets[wallet].set_request_status(amount, PR_PAID, tx_hash=tx_hash)
-        await self.trigger_event(
-            {"event": "new_payment", "address": req.address, "status": req.status, "status_str": req.status_str}, wallet
-        )
+        try:
+            req = self.wallets[wallet].set_request_status(amount, PR_PAID, tx_hash=tx_hash)
+            await self.trigger_event(
+                {"event": "new_payment", "address": req.address, "status": req.status, "status_str": req.status_str}, wallet
+            )
+        except Exception:
+            if self.VERBOSE:
+                print(f"Error processing successful payment {tx_hash} with {amount}:")
+                print(traceback.format_exc())
 
     async def on_shutdown(self, app):
         self.running = False
